requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
chromadb==0.4.15
sentence-transformers==2.2.2
openai==1.3.5
PyPDF2==3.0.1
python-dotenv==1.0.0
pydantic==2.4.2
